app/schemas/parent.py
from pydantic import BaseModel, field_validator


class VerifyPinRequest(BaseModel):
    pin: str

    @field_validator("pin")
    @classmethod
    def pin_must_be_4_digits(cls, v: str) -> str:
        if not v.isdigit() or len(v) != 4:
            raise ValueError("PIN must be exactly 4 digits")
        return v


class ParentTokenResponse(BaseModel):
    token: str
