pyproject.toml
[project]
name = "smarty-steps-backend"
version = "0.1.0"
description = "Smarty Steps learning app backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.3.0",
    "bcrypt>=4.1.0",
    "httpx>=0.27.0",
    "anthropic>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.34.0",
    "uuid7>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "testcontainers[postgres]>=4.8.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
